Refactor: Rename package name to `com.mshdabiola.kotlinmultiplatformtemplate`
This commit renames the project's package name from `com.mshdabiola.kotlin_multiplatform_template` to `com.mshdabiola.kotlinmultiplatformtemplate`.

This change affects various files, including:
- Gradle build files (`app/build.gradle.kts`)
- AndroidManifest related files (badging text files, `google-services.json`)
- Kotlin source files across different source sets (android, common, jvm, wasmJs, desktop) by updating package declarations and import statements.
- Test files by updating package declarations and import statements.

The primary motivation is to remove underscores from the package name, aligning with common Android package naming conventions.
