+++ app/fossReliantRelease-badging.txt
@@ -1,7 +1,7 @@
-package: name='com.mshdabiola.kotlin_multiplatform_template.foss' versionCode='129' versionName='1.2.9' platformBuildVersionName='15' platformBuildVersionCode='35' compileSdkVersion='35' compileSdkVersionCodename='15'
+package: name='com.mshdabiola.kotlinmultiplatformtemplate.foss' versionCode='129' versionName='1.2.9' platformBuildVersionName='15' platformBuildVersionCode='35' compileSdkVersion='35' compileSdkVersionCodename='15'
 minSdkVersion:'26'
 targetSdkVersion:'35'
-uses-permission: name='com.mshdabiola.kotlin_multiplatform_template.foss.DYNAMIC_RECEIVER_NOT_EXPORTED_PERMISSION'
+uses-permission: name='com.mshdabiola.kotlinmultiplatformtemplate.foss.DYNAMIC_RECEIVER_NOT_EXPORTED_PERMISSION'
 application-label:'KotlinMultiplatformTemplate'
 application-label-af:'KotlinMultiplatformTemplate'
 application-label-am:'KotlinMultiplatformTemplate'
@@ -95,7 +95,7 @@ application-icon-480:'res/mipmap-anydpi-v26/ic_launcher.xml'
 application-icon-640:'res/mipmap-anydpi-v26/ic_launcher.xml'
 application-icon-65534:'res/mipmap-anydpi-v26/ic_launcher.xml'
 application: label='KotlinMultiplatformTemplate' icon='res/mipmap-anydpi-v26/ic_launcher.xml'
-launchable-activity: name='com.mshdabiola.kotlin_multiplatform_template.MainActivity'  label='KotlinMultiplatformTemplate' icon=''
+launchable-activity: name='com.mshdabiola.kotlinmultiplatformtemplate.MainActivity'  label='KotlinMultiplatformTemplate' icon=''
 uses-library-not-required:'androidx.window.extensions'
 uses-library-not-required:'androidx.window.sidecar'
 feature-group: label=''
