+++ app/release-badging.txt
@@ -1,4 +1,4 @@
-package: name='com.mshdabiola.kotlin_multiplatform_template' versionCode='3' versionName='1.1.3' platformBuildVersionName='14' platformBuildVersionCode='34' compileSdkVersion='34' compileSdkVersionCodename='14'
+package: name='com.mshdabiola.kotlinmultiplatformtemplate' versionCode='3' versionName='1.1.3' platformBuildVersionName='14' platformBuildVersionCode='34' compileSdkVersion='34' compileSdkVersionCodename='14'
 sdkVersion:'24'
 targetSdkVersion:'34'
 uses-permission: name='android.permission.INTERNET'
@@ -11,7 +11,7 @@ uses-permission: name='android.permission.ACCESS_ADSERVICES_ATTRIBUTION'
 uses-permission: name='android.permission.ACCESS_ADSERVICES_AD_ID'
 uses-permission: name='com.google.android.providers.gsf.permission.READ_GSERVICES'
 uses-permission: name='com.google.android.finsky.permission.BIND_GET_INSTALL_REFERRER_SERVICE'
-uses-permission: name='com.mshdabiola.kotlin_multiplatform_template.DYNAMIC_RECEIVER_NOT_EXPORTED_PERMISSION'
+uses-permission: name='com.mshdabiola.kotlinmultiplatformtemplate.DYNAMIC_RECEIVER_NOT_EXPORTED_PERMISSION'
 application-label:'SkeletonAndroid'
 application-label-af:'SkeletonAndroid'
 application-label-am:'SkeletonAndroid'
@@ -106,7 +106,7 @@ application-icon-480:'res/mipmap-anydpi-v26/ic_launcher.xml'
 application-icon-640:'res/mipmap-anydpi-v26/ic_launcher.xml'
 application-icon-65534:'res/mipmap-anydpi-v26/ic_launcher.xml'
 application: label='SkeletonAndroid' icon='res/mipmap-anydpi-v26/ic_launcher.xml'
-launchable-activity: name='com.mshdabiola.kotlin_multiplatform_template.MainActivity'  label='SkeletonAndroid' icon=''
+launchable-activity: name='com.mshdabiola.kotlinmultiplatformtemplate.MainActivity'  label='SkeletonAndroid' icon=''
 property: name='android.adservices.AD_SERVICES_CONFIG' resource='res/xml/ga_ad_services_config.xml'
 uses-library-not-required:'android.ext.adservices'
 uses-library-not-required:'androidx.window.extensions'
