+++ app/src/commonMain/kotlin/com/mshdabiola/kotlinmultiplatformtemplate/navigation/HydraulicAppNavHost.kt
@@ -15,7 +15,7 @@
  *
  * SPDX-License-Identifier: AGPL-3.0-only
  */
-package com.mshdabiola.kotlin_multiplatform_template.navigation
+package com.mshdabiola.kotlinmultiplatformtemplate.navigation
 
 import androidx.compose.animation.ExperimentalSharedTransitionApi
 import androidx.compose.animation.SharedTransitionLayout
@@ -25,7 +25,7 @@ import androidx.navigation.compose.NavHost
 import com.mshdabiola.detail.navigation.Detail
 import com.mshdabiola.detail.navigation.detailScreen
 import com.mshdabiola.detail.navigation.navigateToDetail
-import com.mshdabiola.kotlin_multiplatform_template.ui.HydraulicAppState
+import com.mshdabiola.kotlinmultiplatformtemplate.ui.HydraulicAppState
 import com.mshdabiola.main.navigation.Main
 import com.mshdabiola.main.navigation.mainScreen
 import com.mshdabiola.setting.navigation.settingScreen
