+++ app/src/jvmMain/kotlin/com/mshdabiola/kotlinmultiplatformtemplate/Writer.kt
@@ -15,7 +15,7 @@
  *
  * SPDX-License-Identifier: AGPL-3.0-only
  */
-package com.mshdabiola.kotlin_multiplatform_template
+package com.mshdabiola.kotlinmultiplatformtemplate
 
 import co.touchlab.kermit.DefaultFormatter
 import co.touchlab.kermit.LogWriter
