+++ app/src/jvmTest/kotlin/com/mshdabiola/kotlinmultiplatformtemplate/HydraulicAppStateTest.kt
@@ -15,7 +15,7 @@
  *
  * SPDX-License-Identifier: AGPL-3.0-only
  */
-package com.mshdabiola.kotlin_multiplatform_template
+package com.mshdabiola.kotlinmultiplatformtemplate
 
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.remember
@@ -24,7 +24,7 @@ import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.createGraph
 import androidx.window.core.layout.WindowSizeClass
-import com.mshdabiola.kotlin_multiplatform_template.ui.HydraulicAppState
+import com.mshdabiola.kotlinmultiplatformtemplate.ui.HydraulicAppState
 import kotlinx.coroutines.test.runTest
 import org.junit.Rule
 import org.junit.Test
