+++ benchmarks/src/main/kotlin/com/mshdabiola/benchmarks/Utils.kt
@@ -29,7 +29,7 @@ import java.io.ByteArrayOutputStream
  */
 val PACKAGE_NAME =
     buildString {
-        append("com.mshdabiola.kotlin_multiplatform_template")
+        append("com.mshdabiola.kotlinmultiplatformtemplate")
         append(BuildConfig.APP_FLAVOR_SUFFIX)
         append(BuildConfig.APP_BUILD_TYPE_SUFFIX)
     }
