+++ conveyor.conf
@@ -26,7 +26,7 @@ app {
   contact-email = "mshdabiola@gmail.com"
 
   // A global identifier for the app.
-  rdns-name = "com.mshdabiola.kotlin_multiplatform_template"
+  rdns-name = "com.mshdabiola.kotlinmultiplatformtemplate"
 
   // Use the Compose logo as an icon and include it so we can pass it to the Window constructor.
   //icons = icons/icon.svg
