+++ instuctions
@@ -18,8 +18,8 @@ export TOKEN="your_actual_token"
     conveyor make site
     ./gradlew run
 
-./gradlew jvmRunHot -PmainClass=com.mshdabiola.kotlin_multiplatform_template.MainAppKt
-./gradlew app:jvmRun -PmainClass=com.mshdabiola.kotlin_multiplatform_template.MainAppKt
+./gradlew jvmRunHot -PmainClass=com.mshdabiola.kotlinmultiplatformtemplate.MainAppKt
+./gradlew app:jvmRun -PmainClass=com.mshdabiola.kotlinmultiplatformtemplate.MainAppKt
 
 //./gradlew assembleDebug -PenableComposeCompilerReports=true
 // ./gradlew assembleDebug -PenableComposeCompilerMetrics=true
