+++ modules/designsystem/.editorconfig
@@ -1,7 +1,7 @@
 
 
 [*.{kt,kts}]
-ktlint_kotlin-multiplatform-template_prefer-method-name=disabled
+ktlint_kotlinmultiplatformtemplate_prefer-method-name=disabled
 
 [*.{kt,kts}]
-ktlint_kotlin-multiplatform-template_prefer-receiver-name=disabled
+ktlint_kotlinmultiplatformtemplate_prefer-receiver-name=disabled
